refactor: src/contents로 Header.tsx 이동
